+++ CHANGES.rst
@@ -18,6 +18,8 @@ Fixed:
 
 * Regression in handling images that don't fit on the current page (issue #153)
 * Fix crash when rendering local table of contents (issue #160)
+* rinoh: variables set in a template configuration file are sometimes ignored
+  (issue #164)
 
 
 Release 0.4.0 (2020-03-05)